+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- `download_file` ([#122](https://github.com/stac-utils/stac-asset/pull/172))
+
 ## [0.3.2] - 2024-05-20
 
 ### Added