+++ CHANGELOG.md
@@ -6,10 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.3.3] - 2024-05-28
+
 ### Added
 
 - `download_file` ([#122](https://github.com/stac-utils/stac-asset/pull/172))
-- allow `text/xml` and `application/xml` to be used interchangeably ([#173](https://github.com/stac-utils/stac-asset/pull/173)) 
+- allow `text/xml` and `application/xml` to be used interchangeably ([#173](https://github.com/stac-utils/stac-asset/pull/173))
 
 ## [0.3.2] - 2024-05-20
 
@@ -149,7 +151,8 @@ Used to be <https://github.com/gadomski/stac-asset>.
 
 Initial release.
 
-[unreleased]: https://github.com/stac-utils/stac-asset/compare/v0.3.2...HEAD
+[unreleased]: https://github.com/stac-utils/stac-asset/compare/v0.3.3...HEAD
+[0.3.3]: <https://github.com/stac-utils/stac-asset/compare/v0.3.2...v0.3.3>
 [0.3.2]: <https://github.com/stac-utils/stac-asset/compare/v0.3.1...v0.3.2>
 [0.3.1]: <https://github.com/stac-utils/stac-asset/compare/v0.3.0...v0.3.1>
 [0.3.0]: <https://github.com/stac-utils/stac-asset/compare/v0.2.4...v0.3.0>